requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.5.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
boto3==1.17.24
botocore==1.20.24; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
click==7.1.2
dnspython==2.1.0; python_version >= '3.6'
email-validator==1.1.2
faker==6.5.2
flask-cors==3.0.8
flask-jwt-extended==3.24.1
flask-login==0.5.0
flask-migrate==2.7.0
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask==1.1.2
gunicorn==20.0.4
idna==3.1; python_version >= '3.4'
itsdangerous==1.1.0
jinja2==2.11.2
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
mako==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==1.1.1
psycopg2-binary==2.8.6
pyjwt==2.0.1; python_version >= '3.6'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.14.0
python-editor==1.0.4
s3transfer==0.3.4
six==1.15.0
sqlalchemy==1.3.19
text-unidecode==1.3
urllib3==1.26.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1
wtforms==2.3.3
